requirements.txt
aioboto3==13.2.0
boto3==1.35.36
fastapi==0.115.6
pydantic==2.10.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
starlette==0.41.3
uvicorn==0.32.1
python-multipart==0.0.19
pymysql==1.1.1
itsdangerous==2.2.0
python-logstash-async==3.0.0
python-logstash==0.4.8